Allow delete of definitive mutation when project participant is not in a definitive revenue distribution yet.

+++ app/Http/Controllers/Api/ParticipantMutation/ParticipantMutationController.php
@@ -121,18 +121,22 @@ class ParticipantMutationController extends ApiController
         DB::transaction(function () use ($participantMutation) {
             $participantProject = $participantMutation->participation;
 
-            $statusLogs = $participantMutation->statusLog;
-            foreach ($statusLogs as $statusLog)
+            if( !$participantProject->participantInDefinitiveRevenue )
             {
-                $statusLog->delete();
+                $statusLogs = $participantMutation->statusLog;
+                foreach ($statusLogs as $statusLog)
+                {
+                    $statusLog->delete();
+                }
+                $participantMutation->delete();
+
+                // Herbereken de afhankelijke gegevens op het participantProject
+                $participantProject->calculator()->run()->save();
+
+                // Herbereken de afhankelijke gegevens op het project
+                $participantProject->project->calculator()->run()->save();
             }
-            $participantMutation->delete();
 
-            // Herbereken de afhankelijke gegevens op het participantProject
-            $participantProject->calculator()->run()->save();
-
-            // Herbereken de afhankelijke gegevens op het project
-            $participantProject->project->calculator()->run()->save();
         });
 
     }
